blaze: kv: store metadata
Store (optional) key metadata in the k-v store, which allows us to
recover the original key and any additional information we want to
attach to it.

fdisk: refactor how partitions work by creating a .blaze directory in
the partition path and use the k-v metadata to indicate that a key is a
partition key in the global k-v store.